[#4] Refactor: Order Table에서 Table Head 분리 및 column 상수화
